2.1.8 - Use AniList as Streaming Links data provider
Since because.moe, the data source that provide links to show where to view titles legally is not being updated, we decided to use AniList to provide this information. As a result, it's no longer possible to set the streaming site region, thus the user have to find out themselves if the title is viewable in the region they are in.